Noti fixbug v1

+++ app/src/main/java/com/example/cs551fitnessapp/database/UserDao.kt
@@ -18,7 +18,7 @@ interface UserDao {
     suspend fun deleteUser(user: UserEntity)
 
     @Query("SELECT * FROM users WHERE userId = :id")
-    suspend fun getUserById(id: Long): UserEntity?
+    suspend fun getUserById(id: Int): UserEntity?
 
     @Query("SELECT * FROM users")
     fun getAllUsers(): Flow<List<UserEntity>>
